Updating awardReducer test to consolidate to one file and removing the extra file
